requirements.txt
PyMuPDF==1.19.1
pdf2image==1.16.0
paddleocr==2.4.0
python-docx==0.8.11
numpy==1.21.0